[BUG] fmbox loses first header for EML files
fmbox was treating eml files like mbox files, assuming the first line was a "From" header, but this is not always the case. For example, Mailcow uses dovecotadm to export Maildir formatted spools that do not have From: as the first line of headers and that header was getting lost.

This change works around the issue by rewinding to the beginning of the file if the first line does not start with "From:".